+++ frontend/components/Description.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+const Description = () => {
+  return (
+    <div className="flex flex-col w-2/5 h-96 items-center justify-center m-10 pb-10">
+      <h1 className="text-lg text-white font-bold">Buy your universal name</h1>
+      <p className="p-4 text-white">
+        Register now your universal name on every layer 2's currently available.
+        No more different identity. Welcome to a cross-chain universe.{" "}
+      </p>
+    </div>
+  );
+};
+
+export default Description;
