+++ frontend/components/Layout.tsx
@@ -1,32 +1,14 @@
 import React, { PropsWithChildren } from "react";
-import { ConnectButton } from "@rainbow-me/rainbowkit";
-import headerImg from "../assets/astro.svg";
-import Image from "next/image";
-
 import "animate.css";
+import Navbar from "./Navbar";
+import Footer from "./Footer";
 
 const Layout = ({ children }: PropsWithChildren) => {
   return (
     <div className="h-screen bg-hero">
-      <div className="flex justify-end p-4">
-        <ConnectButton />
-      </div>
-
+      <Navbar />
       {children}
-      <div className="-translate-y-1/3">
-        <div className={"  flex justify-end animate-waving-hand   "}>
-          <Image
-            className={""}
-            src={headerImg}
-            alt="Header Img"
-            width={300}
-            height={300}
-          />
-        </div>
-      </div>
-      <p className="absolute bottom-0 text-center text-white w-full">
-        Made with ❤️ by your frens at Blockchain Partner
-      </p>
+      <Footer />
     </div>
   );
 };
