+++ frontend/components/ProjectSection.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import RegistrationCard from "./RegistrationCard";
+import Description from "./Description";
+const ProjectSection = () => {
+  return (
+    <div className="flex justify-center items-center mt-20 text-lg">
+      <RegistrationCard />
+      <Description />
+    </div>
+  );
+};
+
+export default ProjectSection;
