+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "dapp_test",
+  "name": "universum",
   "private": true,
   "version": "0.1.0",
   "scripts": {
